refactor(frontend): the lists, their controls and the detail page out of the panel
Three more regions through the `PanelHost` seam, in the pattern the history and
chip regions set: free functions over the host, module doc comments, a `wireXxx`
per region called from `_hydrate`.

- `panel-controls.ts` — the control row above every list (scope pills, the
  saved-Profile picker, Group by, the appliance flat/tree switch, Add/Export)
  plus the grouping glue that turns those choices into buckets. `seg`,
  `menuControl` and `scopeMatches` take no host: they are pure.
- `panel-lists.ts` — the task and appliance lists and the cards they are made
  of, the intro and orphan banners, and the tree layout.
- `panel-detail.ts` — the detail page, the appliance sub-tab sections, and the
  row/link/id primitives. `wireDetail` owns the whole detail branch of
  `_hydrate` including its early return, which it reports as its result, so the
  rule that a task page stops there — and an appliance page does not, or
  `.detail-open` is wired twice — stays in one place.

Emitted HTML is unchanged: every template literal in the built bundle is
byte-identical to the one before (292 of them, compared through the AST).
